Model Test
Changed the test for models from ApiControllerDocTesterTest to ModelTest
